Fixes in the AudioBookPlayback mode
When we didn't know the audiobook yet, the player started at playing
track number 0. However, usually tracks start with number 1.

Also add support to "forget" progress of an audiobook again. The
possible ways to forget the progress are finishing the folder or
stopping the playback (not pausing). In these cases the progress will be
"forgotten".

Also rename and refactor some functions while working on this code.