Add support for sleep during playback mode.
Consider the following scenario: Your kid likes listening to music while
falling asleep. So you turn on the TonuINO for your kid when it goes to
bed, but you don't want the machine to play for ever or until the
current folder finishes but instead only for a couple of minutes. This
is what this feature is meant for. You start the playback and then hold
the sleep-card to the player. The player will continue playing is usual,
but will pause playback after a predefined amount of time. If your kid
wants to continue listening, it can again press the play button and
playback will continue as usual and the timer will be reset.